databricks-f1-intelligence/streamlit_app/requirements.txt
streamlit>=1.30
databricks-sql-connector>=3.0
databricks-sdk>=0.20
plotly>=5.18
pandas>=2.0
pyarrow>=14.0
